Move methods to ndoe class
Make find_all_nodes, find_nodes_in_groups, find_single_nodes &
expand_asterisks class methods of the node class & change refrences to
them accordingly
